+++ packages/vite-plugin-aem-resources/CHANGELOG.md
@@ -6,6 +6,11 @@ documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0] - 2026-06-26
+
+### Changed
+- Version synchronised with the rest of the `@aemvite/*` monorepo. No functional changes in this package.
+
 ## [0.5.1] - 2026-06-26
 
 ### Changed